+++ src/main/java/lotto/domain/RandomNumber.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
